+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/ProjectedStockServiceImpl.java
@@ -32,11 +32,14 @@ import com.axelor.inject.Beans;
 import com.axelor.rpc.Context;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 public class ProjectedStockServiceImpl implements ProjectedStockService {
 
@@ -50,9 +53,10 @@ public class ProjectedStockServiceImpl implements ProjectedStockService {
     Company company = Beans.get(CompanyRepository.class).find(companyId);
     StockLocation stockLocation = stockLocationRepository.find(stockLocationId);
     Mrp mrp = new Mrp();
-    mrp.setStockLocation(findStockLocation(company, stockLocation));
+    stockLocation = findStockLocation(company, stockLocation);
+    mrp.addStockLocationSetItem(stockLocation);
     // If a company has no stockLocation
-    if (mrp.getStockLocation() == null) {
+    if (mrp.getStockLocationSet() == null) {
       return Collections.emptyList();
     }
     mrp.addProductSetItem(product);